rename devices to providers module
