+++ docs/sdk/devices.rst
@@ -6,7 +6,7 @@ Devices
 In this module, you will learn how to use the qBraid SDK to interface with
 quantum backends. We will demonstrate how to construct queries and search
 for available devices using the ``qbraid.get_devices`` function, and
-overview how to execute circuits using the ``qbraid.devices`` module.
+overview how to execute circuits using the ``qbraid.providers`` module.
 
 Unified Device Search
 ----------------------
@@ -106,7 +106,7 @@ how this value is used to wrap the quantum backends / device objects of various 
 Device Wrapper
 ----------------
 
-Given a ``qbraid_id`` retrieved from ``get_devices``, a ``qbraid.devices.DeviceLikeWrapper``
+Given a ``qbraid_id`` retrieved from ``get_devices``, a ``qbraid.providers.DeviceLikeWrapper``
 object can be created as follows:
 
 .. code-block:: python
@@ -192,7 +192,7 @@ least number of queued quantum jobs.
 
 .. code-block:: python
 
-    >>> from qbraid.devices.ibm import ibm_least_busy_qpu
+    >>> from qbraid.providers.ibm import ibm_least_busy_qpu
     >>> qbraid_id = ibm_least_busy_qpu()
     >>> qdevice = device_wrapper(qbraid_id)
     >>> qdevice.name
