+++ qbraid/providers/aws/__init__.py
@@ -13,10 +13,10 @@
 
 """
 ==================================================
-AWS Devices Interface (:mod:`qbraid.devices.aws`)
+AWS Devices Interface (:mod:`qbraid.providers.aws`)
 ==================================================
 
-.. currentmodule:: qbraid.devices.aws
+.. currentmodule:: qbraid.providers.aws
 
 This module contains the classes used to run quantum circuits on devices available through AWS.
