+++ qbraid/providers/aws/job.py
@@ -16,9 +16,9 @@ import logging
 
 from braket.aws import AwsQuantumTask
 
-from qbraid.devices.enums import JOB_FINAL
-from qbraid.devices.exceptions import JobStateError
-from qbraid.devices.job import JobLikeWrapper
+from qbraid.providers.enums import JOB_FINAL
+from qbraid.providers.exceptions import JobStateError
+from qbraid.providers.job import JobLikeWrapper
 
 from .result import AwsGateModelResultWrapper
