+++ qbraid/providers/aws/result.py
@@ -14,7 +14,7 @@ Module defining BraketResultWrapper Class
 """
 import numpy as np
 
-from qbraid.devices.result import ResultWrapper
+from qbraid.providers.result import ResultWrapper
 
 
 class AwsGateModelResultWrapper(ResultWrapper):
