+++ qbraid/providers/result.py
