+++ setup.cfg
@@ -62,8 +62,8 @@ qbraid.transpiler =
     pytket = qbraid.transpiler:PytketCircuitWrapper
     qasm2 = qbraid.transpiler:QasmCircuitWrapper
     qasm3 = qbraid.transpiler:Qasm3CircuitWrapper
-qbraid.devices =
-    aws.device = qbraid.devices.aws:AwsDeviceWrapper
-    aws.job = qbraid.devices.aws:AwsQuantumTaskWrapper
-    ibm.device = qbraid.devices.ibm:IBMBackendWrapper
-    ibm.job = qbraid.devices.ibm:IBMJobWrapper
+qbraid.providers =
+    aws.device = qbraid.providers.aws:AwsDeviceWrapper
+    aws.job = qbraid.providers.aws:AwsQuantumTaskWrapper
+    ibm.device = qbraid.providers.ibm:IBMBackendWrapper
+    ibm.job = qbraid.providers.ibm:IBMJobWrapper
