+++ tox.ini
@@ -30,7 +30,7 @@ commands =
                            qbraid/api \
                            qbraid/interface \
                            qbraid/transpiler \
-                           qbraid/devices \
+                           qbraid/providers \
                            qbraid/tests \
                            -W ignore::DeprecationWarning \
                            -W ignore::PendingDeprecationWarning \
@@ -55,7 +55,7 @@ commands =
     isort . {posargs} qbraid/api \               
                       qbraid/interface \
                       qbraid/transpiler \
-                      qbraid/devices \
+                      qbraid/providers \
                       tools
 
 [testenv:pylint]
@@ -75,7 +75,7 @@ commands =
                      qbraid/api \
                      qbraid/interface \
                      qbraid/transpiler \
-                     qbraid/devices \
+                     qbraid/providers \
                      tools \
                      --disable=C0103,E0401,R0801,R0902,R0903,R0911,R0912,R0914,W0212,W0511
